refactor: Move the useFormData hook to the FormDataProvider file
