--- src/components/Hooks/useFormData.jsx
@@ -1,13 +1,0 @@
-import { useContext } from 'react'
-import FormDataContext from 'src/components/Contexts/FormDataProvider'
-
-/**
- * @returns {import('../../types').FormDataContext}
- */
-export const useFormData = () => {
-  const formDataContext = useContext(FormDataContext)
-  if (!formDataContext) {
-    throw new Error('useFormData must be used within a FormDataProvider')
-  }
-  return formDataContext
-}
