+++ src/components/ModelSteps/ContactDialog.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
 import CompositeHeader from 'src/components/CompositeHeader/CompositeHeader'
-import { useFormData } from 'src/components/Hooks/useFormData'
+import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import ContactList from 'src/components/ModelSteps/ContactList'
 import SubmitButton from 'src/components/ModelSteps/widgets/SubmitButton'
