+++ src/components/ModelSteps/ContactDialog.spec.jsx
@@ -3,8 +3,8 @@ import '@testing-library/jest-dom'
 import { render, waitFor } from '@testing-library/react'
 import React from 'react'
 import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
+import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { StepperDialogProvider } from 'src/components/Contexts/StepperDialogProvider'
-import { useFormData } from 'src/components/Hooks/useFormData'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import ContactDialog from 'src/components/ModelSteps/ContactDialog'
 import { fetchCurrentUser } from 'src/helpers/fetchCurrentUser'
