+++ src/components/ModelSteps/InformationDialog.jsx
@@ -6,7 +6,7 @@ import React, { useCallback, useMemo, useState } from 'react'
 import IlluGenericInputDate from 'src/assets/icons/IlluGenericInputDate.svg'
 import IlluGenericInputText from 'src/assets/icons/IlluGenericInputText.svg'
 import CompositeHeader from 'src/components/CompositeHeader/CompositeHeader'
-import { useFormData } from 'src/components/Hooks/useFormData'
+import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import { normalizeFormdataMetadata } from 'src/components/ModelSteps/helpers'
 import SubmitButton from 'src/components/ModelSteps/widgets/SubmitButton'
