+++ src/components/ModelSteps/Scan/ScanActions/ScanActionsWrapper.spec.jsx
@@ -2,8 +2,8 @@
 import { render, waitFor } from '@testing-library/react'
 import React from 'react'
 import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
+import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { StepperDialogProvider } from 'src/components/Contexts/StepperDialogProvider'
-import { useFormData } from 'src/components/Hooks/useFormData'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import ScanWrapper from 'src/components/ModelSteps/Scan/ScanWrapper'
 import AppLike from 'test/components/AppLike'
