+++ src/components/ModelSteps/ScanResult/ScanResultCard.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React, { useState, forwardRef } from 'react'
-import { useFormData } from 'src/components/Hooks/useFormData'
+import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import styles from 'src/components/ModelSteps/ScanResult/ScanResultCard.styl'
 import ScanResultCardActions from 'src/components/ModelSteps/ScanResult/ScanResultCardActions'
