+++ src/components/ModelSteps/ScanResult/ScanResultDialog.spec.jsx
@@ -1,8 +1,8 @@
 import { render, fireEvent, waitFor } from '@testing-library/react'
 import React from 'react'
 import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
+import { useFormData } from 'src/components/Contexts/FormDataProvider'
 import { StepperDialogProvider } from 'src/components/Contexts/StepperDialogProvider'
-import { useFormData } from 'src/components/Hooks/useFormData'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import ScanResultDialog from 'src/components/ModelSteps/ScanResult/ScanResultDialog'
 import { getAttributesFromOcr } from 'src/components/ModelSteps/helpers'
